wow that didnt take too long. makes you wonder what i was doing all day

+++ src/map.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <fstream>
 
-int** loadMap(std::string path, unsigned long &mapWidth, unsigned long &mapHeight)
+int** loadMap(std::string path, unsigned int &mapWidth, unsigned int &mapHeight)
 {
     std::ifstream mapFile;
     mapFile.open(path);
